Removed stability check for the outermost level
The code will periodically check to see if there are any unstable triples or
quadruples according to the criterion set out in Mardling & Aarseth (2001).
If there are any unstable hierarchies, the code will keep integrating.  But
if all hierarchies are stable it will stop.

In the earlier version of the code, the stability check would include the
outermost possible hierarchy.  (Thus if there were five stars, the code
would check to see if the outermost quintuple was stable, if it exists.)
This is problematic because in scattering experiments, such systems can
never be stable -- the Mardling criterion will sometimes label them as
stable, however.  This leads to systems halting prematurely.

I removed this check at the outermost level.  Although this is aimed
at scattering experiments, it should benefit experiments of the long-term
evolution of isolated systems.  Since one may be interested in the long-term
evolution of a stable system, one would also not want such systems to halt.